+++ src/main/java/me/noodles/dayandnight/UpdateJoinEvent.java
@@ -1,4 +1,4 @@
-package me.noodles.dayandnight.requested;
+package me.noodles.dayandnight;
 
 import org.bukkit.event.player.*;
 import org.bukkit.ChatColor;
@@ -14,14 +14,14 @@ public class UpdateJoinEvent implements Listener
     public void onJoin(final PlayerJoinEvent e) {
     	Player p = e.getPlayer();
     	if (p.hasPermission("dayandnight.update")) {
-    		if (MainDayNight.getPlugin().getConfig().getBoolean("Update.Enabled") == true){
-    		this.checker = new UpdateChecker(MainDayNight.plugin);
+    		if (DayAndNight.getPlugin().getConfig().getBoolean("Update.Enabled") == true){
+    		this.checker = new UpdateChecker(DayAndNight.plugin);
                         if (this.checker.isConnected()) {
                             if (this.checker.hasUpdate()) {
                             	p.sendMessage(ChatColor.GRAY + "=========================");
                                 p.sendMessage(ChatColor.RED + "DayAndNight is outdated!");
                                 p.sendMessage(ChatColor.GREEN + "Newest version: " + this.checker.getLatestVersion());
-                                p.sendMessage(ChatColor.RED + "Your version: " + MainDayNight.plugin.getDescription().getVersion());
+                                p.sendMessage(ChatColor.RED + "Your version: " + DayAndNight.plugin.getDescription().getVersion());
                                 p.sendMessage(ChatColor.GRAY + "=========================");
                             }
                         }
