+++ src/main/resources/plugin.yml
@@ -1,0 +1,10 @@
+name: DayAndNight
+version: ${project.version}
+main: me.noodles.dayandnight.DayAndNight
+authors: [BGHDDevelopment]
+website: https://bghddevelopment.com/
+commands:
+  day:
+    description: Sets Time To Day
+  night:
+    description: Sets Time To Night
